# api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "OpenPhotobox API",
        "version": "1.0.0",
        "description": "Family photo management system with face recognition, albums, and sharing"
    },
    "paths": {
...
        "/people/": {
            "get": {
                "operationId": "people_list",
                "description": "ViewSet for managing people and face recognition.",
                "summary": "List people",
                "parameters": [
                    {
                        "name": "cursor",
                        "required": false,
                        "in": "query",
                        "description": "The pagination cursor value.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "Number of results to return per page.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "tags": [
                    "People"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedPersonList"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "people_create",
                "description": "ViewSet for managing people and face recognition.",
                "summary": "Create person",
                "tags": [
                    "People"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginatedPersonList": {
                "type": "object",
                "required": [
                    "results"
                ],
                "properties": {
                    "next": {
                        "type": "string",
                        "nullable": true,
                        "format": "uri",
                        "example": "http://api.example.org/accounts/?cursor=cD00ODY%3D\""
                    },
                    "previous": {
                        "type": "string",
                        "nullable": true,
                        "format": "uri",
                        "example": "http://api.example.org/accounts/?cursor=cj0xJnA9NDg3"
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Person"
                        }
                    }
                }
            },
...
            "Person": {
                "type": "object",
                "description": "Serializer for Person model",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "displayName": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "aka": {},
                    "notes": {
                        "type": "string"
                    },
                    "faceCount": {
                        "type": "string",
                        "readOnly": true
                    },
                    "assetCount": {
                        "type": "string",
                        "readOnly": true
                    },
                    "embeddingCentroid": {
                        "type": "string",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "displayName"
                ]
            },
...
            "PersonRequest": {
                "type": "object",
                "description": "Serializer for Person model",
                "properties": {
                    "displayName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 255
                    },
                    "aka": {},
                    "notes": {
                        "type": "string"
                    },
                    "embeddingCentroid": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "displayName"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "tokenAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Token-based authentication with required prefix \"Token\""
            }
        }
    },
...
}